requirements.txt
django>=5.0
requests
aiohttp
beautifulsoup4
colorama
dnspython
pyyaml
